requirements.txt
# NEO: Neural Executive Operator - Dependencies

# Core AI and Machine Learning
torch>=2.0.0
tensorflow>=2.13.0
transformers>=4.30.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Natural Language Processing
nltk>=3.8.0
spacy>=3.6.0
langchain>=0.0.300
openai>=1.0.0
anthropic>=0.3.0
google-generativeai>=0.3.0

# Deep Learning and Neural Networks
keras>=2.13.0
pytorch-lightning>=2.0.0

# Computer Vision (if needed)
opencv-python>=4.8.0
pillow>=10.0.0

# System Control and Automation
psutil>=5.9.0
pyautogui>=0.9.54
schedule>=1.2.0
python-crontab>=3.0.0

# Cybersecurity and Penetration Testing
scapy>=2.5.0
#nmap>=0.0.1
python-nmap>=0.7.1
cryptography>=41.0.0
paramiko>=3.3.0

# Code Analysis and Development
pylint>=2.17.0
black>=23.7.0
autopep8>=2.0.4
rope>=1.9.0
jedi>=0.19.0
radon>=6.0.1

# Web and API
fastapi>=0.103.0
uvicorn>=0.23.0
requests>=2.31.0
aiohttp>=3.8.5
beautifulsoup4>=4.12.0

# Database
sqlalchemy>=2.0.0
redis>=5.0.0
pymongo>=4.5.0

# Research and Data Collection
wikipedia>=1.4.0
arxiv>=1.4.8
scholarly>=1.7.11

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
click>=8.1.7
rich>=13.5.0
colorama>=0.4.6
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Logging and Monitoring
loguru>=0.7.0
prometheus-client>=0.17.1

# Task Queue
celery>=5.3.0
rabbitmq>=0.2.0

# Security
pyjwt>=2.8.0
bcrypt>=4.0.1

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0
